Cleaned up tag encoding, now with clear chunk field
Before, the tag format's type field was limited to 9-bits. This sounds
like a lot, but this field needed to encode up to 256 user-specified
types. This limited the flexibility of the encoded types. As time went
on, more bits in the type field were repurposed for various things,
leaving a rather fragile type field.

Here we make the jump to full 12-bit type fields. This comes at the cost
of a smaller length field, however the use of the length field was
always going to come with a RAM limitation. Rather than putting pressure
on RAM for inline files, the new type field lets us encode a chunk
number, splitting up inline files into multiple updatable units. This
actually pushes the theoretical inline max from 8KiB to 256KiB! (Note
that we only allow a single 1KiB chunk for now, chunky inline files
is just a theoretical future improvement).

Here is the new 32-bit tag format, note that there are three levels
of types which break down into more info:

[----                 32                 ----]
[1|---    12     ---|--   9   --|--   10   --]
 ^.        ^        .     ^            ^- entry length
 |.        |        .     \-------------- file id chunk info
 |.        \--------.-------------------- type info (type3)
 \.-----------------.-------------------- valid bit
  [-  4  -|--  8  --]
  .   ^   .    ^- chunk info
  .   \---.------ specific type (type2)
  [- 3 -|1]
     ^   ^------- modifier bit
     \----------- generic type (type1)

Additionally, I've split the CREATE tag into separate SPLICE and FILE
tags. This simplified the new compact logic a bit. For now, littlefs
still follows the rule that a FILE tag precedes any other tags related
to a file, but this can change in the future.